Count discarded incoming messages
An incoming Diameter message is either a request, an answer to an
outstanding request, or an unexpected answer. The latter weren't
counted, but are now counted on keys of this form:

  {pid(), {{unknown, 0}, recv, discarded}}

The form of the second element is similar to those of other counters,
like:

  {{relay, 0|1}, send|recv, invalid_error_bit}

Compare this to the key used when counting known answers:

  {{ApplicationId, CommandCode, 0}, recv}

The application id and command code aren't included so as not to count
on arbitrary keys, a topic last visited in commit 49e8b11c.