inets: Add profile option
To enable the HTTP server to run in a virtualized environment,
where there can be more that one server that has the same
bind_address and port, we add a new option profile.
The profile name will be used in concatenation with bind_address and port
to identify the HTTP server instance.

The name profile was chosen as there is a similar concept in the HTTP client
where profile names can be used to instantiate client configurations.