Change hipe_bifs:system_crc/1 to hipe_bifs:system_crc/0
The macro HIPE_SYSTEM_CRC used to contain a hidden cookie from the VM
that generated hipe_literals.hrl.  This means that BEAM files containing
that macro would be tied to a particular version of the VM.

Change hipe_bifs:system_crc such that it doesn't require a hidden cookie
to return the desired value.