Annotate used types in erl_lint
This is a follow up to commit b5ee5c6.

Similar to function calls, we also need to annotate used types
on call site since this information is used later on when showing
undefined type errors.

Notice we don't need to annotate the type export because
it is an attribute and those are always properly annotated.