Clean up rpc_SUITE:call_benchmark/1
The call_benchmark/1 test case is supposed to return a comment
about the number of RPC calls per second, but the return value
from the function that produces the comment (do_call_benchmark/4)
is ignored. The test case is essentially useless except as a
smoke test.

Simplify the test case and correct the showing of the number of
RPC calls per second, as well as eliminate the use of erlang:now/0.