Speed up supervisor:count_children/1; RT=non-temporary
Speed up supervisor:count_children/1 for simple_one_for_one supervisors
with 'permanent' or 'transient' children. This is achieved by avoiding
looping through all the child process and verifying that each one is
alive. For a supervisor with 100,000 processes the count-time will drop
from approx 30ms to about 0.005ms. This avoids having the supervisor
block for an extended period while the count takes place. Under normal
circumstances the accuracy of the result should also improve since the
duration is too short for many processes to die during the count.