ssh: Option unexpectedfun for ssh:daemon and ssh:connect
This option has a fun as value.  The fun will be called when an
unexpected message arrives.  The fun returns either 'skip' or 'report'
to guide the connection_handler what to do.

One usage is to filter out messages that are not wanted in the error
logger as info reports. An example of such a message is the 'etimedout'
tcp error message that will be received if a connection has keep_alive
and the peer is restarted.