inets: Remove use of httpd_conf:clean/1 and httpd_conf:custom_clean/3
Internal use of the function white_space_clean/1 could probably be done
in a much better way using re-module and removing a lot of legacy code.
But we will have to do this later, due to lack of time, we want to make
this commit as little work as possible.