Make the 'r' operand type optional
The 'r' type is now mandatory. That means in order to handle
both of the following instructions:

  move x(0) y(7)
  move x(1) y(7)

we would need to define two specific operations in ops.tab:

  move r y
  move x y

We want to make 'r' operands optional. That is, if we have
only this specific instruction:

  move x y

it will match both of the following instructions:

  move x(0) y(7)
  move x(1) y(7)

Make 'r' optional allows us to save code space when we don't
want to make handling of x(0) a special case, but we can still
use 'r' to optimize commonly used instructions.