Introduce specialized versions of move2
Currently, move2/2 does the two moves sequentially to ensure
that the instruction will always work correctly.

We can do better than that. If the two move instructions have
any registers in common, we can introduce simpler and slightly
more efficient instructions to handle those cases:

  move_shift/3
  move_dup/3

For the remaining cases when the the move instructions
have no common registers, the move2/4 instruction can perform
the moves in parallel which is probably slightly more efficient.
For clarity's sake, we will remain the instruction to move2_par/4.