inets: Fix broken fd feature
If a wrapper script is used to open a privileged port before starting
erlang and supplied as a init argument to erlang it should override
the port configured by the httpd config file. Some of the code handling
this was lost, in the commit da3797589463002f28fc42ef27872650fe1de938,
due to missing tests of this functionality.

This commit adds the code for this in a better place.
In 18 also add a test.