Improve unpacking performance on x86_64
When unpacking operands on 64-bit CPUs, use a smarter mask to
help the compiler optimize the code.

It turns out that on x86_64, if we use the mask 0xFFFFUL (selecting
the 16 least significant bits), the compiler can combine a move and
a mask operation into the single insruction 'movzwl', which will
eliminate one instruction.