Don't compute AVP list length unnecessarily at AVP decode
This has had a hugely negative impact on performance when decoding
messages containing many AVP: each decode of an AVP having variable
arity computed the length of the list of previously decoded AVPs when
checking that the allowed arity was not exceeded, even if the allowed
arity was infinite, making for O(n^2) cost. Here are some execution
times, for diameter_codec:decode/2 on a representative message with n
integer AVPs in the Common application (on the host at hand):

             Before      After
             -------     ---------
  n = 1K       5 ms        2 ms
  n = 10K    500 ms       25 ms
  n = 100K    75 sec     225 ms
  n = 1M                   2.6 sec

Note the nearly linear increase following the change.

Remove the dire documentation warning for incoming_maxlen as a
consequence. It can still be useful to set, but not doing so won't have
the same consequences as previously.