Teach smp VM how to deal with crash of a linked trace port
Problem: The sys-msg-dispather crashes the VM when trying to send
exit signals from the links of the terminating trace port.
If try-lock of the linked process fails, a pending exit is
scheduled and erts_scheduler_data() is then called to find
"my" run queue. But sys-msg-dispatcher is not a scheduler
and has no scheduler data, hence SEGV.

Fix: If not a scheduler and we cannot get process locks,
schedule process in its previous run-queue.