Erlang shell: Support keys Del, Home and End
Add support for the Delete, Home and End keys in the Erlang shell.
These keys are ubiquitous on modern keyboards. Having them working
as expected adds to the convenience of working in the shell, since
they are much easier to use than the corresponding Ctrl-{D,A,E} keys.

The implementation is in line with the existing framework of the line
editor and is thus a natural (and minimal) extension.