Set xml namespace to http://www.w3.org/XML/1998/namespace.
Statically set the xml namespace to http://www.w3.org/XML/1998/namespace, as
stated in 'Namespaces in XML 1.1 (Second Edition)' [1]:
"The prefix xml is by definition bound to the namespace name http://www.w3.org/XML/1998/namespace.
It MAY, but need not, be declared, and MUST NOT be undeclared or bound to any other namespace name.
Other prefixes MUST NOT be bound to this namespace name, and it MUST NOT be declared as the
default namespace. "

[1] http://www.w3.org/TR/xml-names11/