erts: fix binary_to_integer boundary case
erlang:binary_to_integer/1 and /2 fail to detect invalid
input consisting of a single + or - sign but nothing else.
For an input like <<"+">> they return 0, while list_to_integer/1
correctly signals a badarg for "+".

Fixed by checking if the input is empty after the initial +/-
sign processing.

Added a test case which fails without this fix but passes with it.

Thanks to "niku" for reporting the issue.