v3_core: Improve code generation for guards
When translating guards to Core Erlang, it is sometimes necessary
to add an is_boolean/1 guard test. Here is an example when it is
necessary:

  o(A, B) when A or B ->
    ok.

That would be translated to something like:

  o(A, B) when ((A =:= true) or (B =:= true)) and
      is_boolean(A) and is_boolean(B) ->
    ok.

The is_boolean/1 tests are necessary to ensure that the guard
fails for calls such as:

  o(true, not_boolean)

However, because of a bug in v3_core, is_boolean/1 tests were
added when they were not necessary. Here is an example:

  f(B) when not B -> ok.

That would be translated to:

  f(B) when (B =:= false) and is_boolean(B) -> ok.

The following translation will work just as well.

  f(B) when B =:= false -> ok.

Correct the bug to suppress those unnecessary is_boolean/1 tests.