erts: Fix uds socket handling for os x
OS X is very strict in what it requires of you and also
gives strange error codes for some errors. In this commit
we remap EMSGSIZE to EMFILE and also precisely tune the
amount of data we send on the socket so that we can recv
only that data. If we try to recv more, recvmsg fails with
EPIPE if the remote end has been closed.