Optimize zip:unzip/2 when uncompressing to memory
Optimize the case where we are appending to the end of the binary, use
binary syntax to create binaries with room for expansion in the next
loop, instead of using iolist_to_binary which creates a binary of the
exact size and needs to be copied in each loop.

Also remove support the unused Acc as iolists.