inets: Add test suite for socket_wrap feature
On Linux systems a socket may be opened pre starting Erlang and
then passed to Erlangs ftpd or httpd daemon. This is useful as the wrap program
can open a privileged port and Erlang does not have to be run as root.

The test program will only open a random port, but it exercises the code
so that the feature will remain intact.