Add test for "Fix cover output file"
If a module includes eunit.hrl, a parse transform adds the function
test/0 on line 0 in the module.  A bug in OTP-18.0 caused
cover:analyse_to_file/1 to fail to insert cover data in the output
file when line 0 existed in the cover data table. The bug is corrected
by the commit "Fix cover output file".

This commit adds a test which checks that the bug is not introduced
again.