v3_codegen: Optimize matching of the final size-less binary segment
Consider the following function:

  f(Bin, Bool) ->
      case Bin of
	<<Val:16/binary,_/binary>> when Bool ->
	  Val
      end.

Simplified, the generated code looks like:

    bs_start_match2 Fail Live Bin => Bin
    bs_get_integer2 Fail Live Bin size=Sz  unit=1 => Val
    bs_skip_bits2   Fail      Bin size=all unit=8
    is_eq_exact Fail Bool true

The code generator will replace the bs_skip_bits2 instruction with
a bs_test_unit instruction if it can be clearly seen that the
context register will not be used again. In this case, it is not
obvious without looking at the code at the Fail label.

However, it turns out that bs_test_unit instruction is always
safe beacuse of the way v3_kernel compiles pattern matching.
It doesn't matter whether the match context will be used again.
If it will be used again, the position in it will *not* be used.
Instead, a bs_restore2 instruction will restore one of the saved
instructions.