Introduce the three stages Root/TrunkDbTerm structs
RootDbTerm can now assume three forms: stage 1, stage 2 and stage 3.
The Least Significant Bit of the pointer to RootDbTerm determines
whether it is a stage 1 or a stage 2/3 RootDbTerm.
The Least Significant Bit of RootDbTerm's dt.tail.trunk member
determines whether it is a stage 2 or stage 3 RootDbTerm.

NOTE: Stage 2 and stage 3 are equivalent to the previous RootDbTerm
      implementations without and with the nested LHT, respectively.