beam_dead: Optimize select_val instructions
In a select_val instruction, values associated with a label
which is the same as the failure label can be removed. We
already do this optimization in beam_clean, but it is better
do this sort of optimization before the beam_jump pass.

Also rewrite a select_val instruction with a single value
to is_eq_exact instruction followed by a jump instruction.