beam_dead: Improve optimization of literal binary matching
When the bit syntax is used to match a single binary literal, the bit
syntax instructions will be replaced with a comparison to a binary
literal. The only problem is that the bs_context_to_binary instruction
will not be eliminated.

Example:

f(<<"string">>) ->
   ok.

This function would be translated to:

{function, f, 1, 2}.
  {label,1}.
    {line,...}.
    {func_info,...}.
  {label,2}.
    {test,is_eq_exact,{f,3},[{x,0},{literal,<<"string">>}]}.
    {move,{atom,ok},{x,0}}.
    return.
  {label,3}.
    {bs_context_to_binary,{x,0}}.
    {jump,{f,1}}.

The bs_context_to_binary instruction serves no useful purpose,
since {x,0} can never be a match context. Eliminating the
instruction, the resulting code will be:

{function, f, 1, 2}.
  {label,1}.
    {line,...}.
    {func_info,...}.
  {label,2}.
    {test,is_eq_exact,{f,1},[{x,0},{literal,<<"string">>}]}.
    {move,{atom,ok},{x,0}}.
    return.