beam_type: Improve optimization by keeping track of integers
The ASN.1 compiler often generates code similar to:

  f(<<0:1,...>>) -> ...;
  f(<<1:1,...>>) -> ....

Internally that will be rewritten to (conceptually):

  f(<<B:1,Tail/binary>>) ->
    case B of
      0 ->
        case Tail of ... end;
      1 ->
        case Tail of ... end;
      _ ->
        error(case_clause)
   end.

Since B comes from a bit field of one bit, we know that the only
possible values are 0 and 1. Therefore the error clause can be
eliminated like this:

  f(<<B:1,Tail/binary>>) ->
    case B of
      0 ->
        case Tail of ... end;
      _ ->
        case Tail of ... end
   end.

Similarly, we can also a deduce the range for an integer from
a 'band' operation with a literal integer.

While we are at it, also add a test case to improve the coverage.