Flush timeout message from message queue when canceling timer
In ct_netconfc, if a timer expired 'at the same time' as the server
sent the rpc-reply, the timeout message might already be in the
client's message queue when the client removed the timer ref from its
'pending' list. This caused a crash in the client since the timer ref
could no longer be found when handling the timeout message.

This commit fixes the problem by always flushing the timeout message
from the message queue when canceling a timer.